// generated/src/app/api-models/cap-entity-select.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class CapEntitySelect {
  capModelId: string;
  capEntityCode: string;
  capEntityId: string;
  seqCapEntitySel: number;
  columnName: string;
  columnOccurrence: number;
  operator: string;
  fromValue: string;
  thruValue: string;
  securityCode: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
  columnType: string;
}